\name{susendal}
\alias{susendal}
\title{Daily Susendal record.}
\description{
  A station record of daily mean temperature and daily precipitation from Susendal. 
}
\usage{data(susendal)}
\format{
The dataset is a data.frame containing:
   \code{V1}:station number;
   \code{V2}:a vector holding day of month;
   \code{V3}:a vector holding the month;
   \code{V4}:a vector holding the year;
   \code{V5}:a vector holding daily precipitation in mm;
   \code{V6}:a vector holding daily mean temperature in deg C;
      }
\source{The Norwegian Meteorological Institute, Climatology division.}
\references{
The Norwegian Meteorological Institute, P.O. Box 43, 0313 Oslo, Norway (URL \url{http://www.met.no})
}
\keyword{datasets}





